# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for F1 Grand-Prix info",
        "title": "FormulOne Web-Server",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5259",
    "basePath": "/",
    "paths": {
...
        "/api/grandprix": {
            "get": {
                "description": "Get all gp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gp"
                ],
                "summary": "Get all gp",
                "operationId": "get-all-gp",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GrandPrix"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "Create gp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gp"
                ],
                "summary": "Create gp",
                "operationId": "create-gp",
                "parameters": [
                    {
                        "description": "GP info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GrandPrix"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/grandprix/season/{season}": {
            "get": {
                "description": "Get gp by season",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gp"
                ],
                "summary": "Get gp by season",
                "operationId": "get-gp-by-season",
                "parameters": [
                    {
                        "type": "string",
                        "description": "season",
                        "name": "season",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GrandPrix"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/grandprix/{gp_id}/qual_results": {
            "post": {
                "description": "Create qual_result",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qual_results"
                ],
                "summary": "Create qual_result",
                "operationId": "create-qual_result",
                "parameters": [
                    {
                        "type": "string",
                        "description": "gp_id",
                        "name": "gp_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "qual result info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.QualResult"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/api/grandprix/{gp_id}/race_results": {
            "post": {
                "description": "Create race_results",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "race_resultss"
                ],
                "summary": "Create race_results",
                "operationId": "create-race_results",
                "parameters": [
                    {
                        "type": "string",
                        "description": "gp_id",
                        "name": "gp_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "race result info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RaceResult"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/api/grandprix/{id}": {
            "get": {
                "description": "Get gp by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gp"
                ],
                "summary": "Get gp by id",
                "operationId": "get-gp-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GrandPrix"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "description": "Update gp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gp"
                ],
                "summary": "Update gp",
                "operationId": "update-gp",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "GP info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GrandPrix"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "delete gp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gp"
                ],
                "summary": "Delete gp",
                "operationId": "delete-gp",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/grandprix/{id}/qual_results": {
            "get": {
                "description": "Get qualresults of gp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gp"
                ],
                "summary": "Get qualresults of gp",
                "operationId": "get-qual-of-gp",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.QualResultView"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/grandprix/{id}/race_results": {
            "get": {
                "description": "Get raceresults of gp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gp"
                ],
                "summary": "Get raceresults of gp",
                "operationId": "get-race-of-gp",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RaceResultView"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/grandprix/{id}/race_winner": {
            "get": {
                "description": "Get race winner of gp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gp"
                ],
                "summary": "Get race winner of gp",
                "operationId": "get-race-winner-of-gp",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RaceResultView"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.GrandPrix": {
            "type": "object",
            "properties": {
                "gp_date_num": {
                    "type": "integer"
                },
                "gp_id": {
                    "type": "integer"
                },
                "gp_month": {
                    "type": "string"
                },
                "gp_name": {
                    "type": "string"
                },
                "gp_place": {
                    "type": "string"
                },
                "gp_season": {
                    "type": "integer"
                },
                "gp_track_id": {
                    "type": "integer"
                }
            }
        },
        "models.QualResult": {
            "type": "object",
            "properties": {
                "gp_id": {
                    "type": "integer"
                },
                "q1_time": {
                    "type": "string"
                },
                "q2_time": {
                    "type": "string"
                },
                "q3_time": {
                    "type": "string"
                },
                "qual_driver_id": {
                    "type": "integer"
                },
                "qual_driver_place": {
                    "type": "integer"
                },
                "qual_id": {
                    "type": "integer"
                },
                "qual_team_id": {
                    "type": "integer"
                }
            }
        },
        "models.QualResultView": {
            "type": "object",
            "properties": {
                "driver_name": {
                    "type": "string"
                },
                "gp_name": {
                    "type": "string"
                },
                "q1_time": {
                    "type": "string"
                },
                "q2_time": {
                    "type": "string"
                },
                "q3_time": {
                    "type": "string"
                },
                "qual_driver_place": {
                    "type": "integer"
                },
                "qual_id": {
                    "type": "integer"
                },
                "team_name": {
                    "type": "string"
                }
            }
        },
        "models.RaceResult": {
            "type": "object",
            "properties": {
                "gp_id": {
                    "type": "integer"
                },
                "race_driver_id": {
                    "type": "integer"
                },
                "race_driver_place": {
                    "type": "integer"
                },
                "race_id": {
                    "type": "integer"
                },
                "race_team_id": {
                    "type": "integer"
                }
            }
        },
        "models.RaceResultView": {
            "type": "object",
            "properties": {
                "driver_name": {
                    "type": "string"
                },
                "gp_name": {
                    "type": "string"
                },
                "race_driver_place": {
                    "type": "integer"
                },
                "race_id": {
                    "type": "integer"
                },
                "team_name": {
                    "type": "string"
                }
            }
        },
...
    }
}